Further code structural improvements * Abstracted helpers in separate folder * Removed useless components * Improved code structure of cart
